Separated the general style parsing code into its own file
Styles specific to tables and normal text are still in those files, so
these are the ones common to both
